32-bit font_trait and setting trait, which increases performance. Single glyph cache key, which simplify the structure of the glyph cache. Glyph run LRU stores the 10 latest accessed strings. Remove database-related code. Database experiments goes to separate branch. Serialize setting trait.

+++ gdimm/ft_renderer.h
@@ -22,7 +22,7 @@ class gdimm_ft_renderer : public gdimm_renderer
 		FT_ULong load_flags,
 		bool is_italic,
 		bool request_outline,
-		uint64_t font_trait) const;
+		unsigned int font_trait) const;
 	bool generate_glyph_run(bool is_glyph_index, LPCWSTR lpString, UINT c, glyph_run &new_glyph_run, bool request_outline);
 
 	bool render(bool is_glyph_index, bool is_pdy, LPCWSTR lpString, UINT c, CONST INT *lpDx, glyph_run &new_glyph_run);
